HUSH-843: UnitCount should be float to handle... (#116)
* HUSH-843: UnitCount should be float to handle...

* HUSH-843: Moved changeset to previous changelog

+++ Application/src/main/java/com/ikea/imc/pam/budget/service/repository/model/Expenses.java
@@ -29,7 +29,7 @@ public class Expenses extends AbstractEntity {
     
     private Byte internalPercent;
     
-    private Short units;
+    private Double units;
     
     private Byte weeks; //TODO Not in use but retained for future in DB
